requirements.txt
praw==7.7.1
tweepy==4.14.0
python-dotenv==1.0.0
APScheduler==3.10.1 